Split issue-tool tests by command family
The worktree issue suite had become a 2,300-line monolith that mixed queue logic, worktree flows, batch launches, summary helpers, and closeout behavior in one file. This change decomposes it into command-family files with one shared support module, which keeps the active package shape visible in the tests and lowers the worst per-test monkeypatch density below the previous red-line.

Constraint: CLI behavior had to stay stable while replacing the monolithic test surface
Constraint: The split had to run against the canonical scripts.issue_tool package path rather than the legacy shim namespace
Rejected: Keep a single test_worktree_issues.py and only add helper fixtures | Left the package/test shape mismatch intact and did not reduce test navigation cost
Rejected: Move all support into conftest.py | Broadened fixture scope beyond the issue-tool slice without need
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Add new issue-tool tests to the command-family files first; do not reintroduce a catch-all test_worktree_issues.py monolith
Tested: uv run pytest tests/unit/test_issue_tool_queue_evidence.py tests/unit/test_issue_tool_worktree_flows.py tests/unit/test_issue_tool_gitnexus_summary.py tests/unit/test_issue_tool_batch.py tests/unit/test_issue_tool_launch_closeout.py tests/unit/test_issue_tool_git_utils.py -q; uv run ruff check tests/unit/issue_tool_test_support.py tests/unit/test_issue_tool_queue_evidence.py tests/unit/test_issue_tool_worktree_flows.py tests/unit/test_issue_tool_gitnexus_summary.py tests/unit/test_issue_tool_batch.py tests/unit/test_issue_tool_launch_closeout.py tests/unit/test_issue_tool_git_utils.py; make validate-local; make pre-validate-session
Not-tested: Whole-repo unit runtime outside the issue-tool-related test slice
